o Fixed bugs. o More work done on staff.
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/qi4j/apps/chronos@8201 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ core/src/main/java/org/qi4j/chronos/model/mixins/MoneyDescriptorMixin.java
@@ -12,7 +12,6 @@
  */
 package org.qi4j.chronos.model.mixins;
 
-import java.math.BigDecimal;
 import java.util.Currency;
 import org.qi4j.api.annotation.Uses;
 import org.qi4j.library.general.model.Descriptor;
@@ -34,7 +33,7 @@ public final class MoneyDescriptorMixin implements Descriptor
             currencyCode = Character.SPACE_SEPARATOR + currency.getCurrencyCode();
         }
 
-        BigDecimal amount = money.getAmount();
+        Integer amount = money.getAmount();
         return amount + currencyCode;
     }
